requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
pydantic>=2.0.0
httpx>=0.27.0
uvicorn>=0.30.0
fastapi>=0.112.0

# DeepSeek LLM support
openai>=1.0.0
langchain-openai>=0.2.0
langchain-deepseek>=0.1.0

# Vector stores and embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Memory and persistence
redis>=5.0.0

# Web search capabilities
duckduckgo-search>=6.0.0
tavily-python>=0.3.0

# MCP support
# mcp>=1.0.0  # 如果MCP包不存在，可以暂时注释掉

# Utilities
python-dotenv>=1.0.0
structlog>=24.0.0
typer>=0.12.0
rich>=13.0.0
